fixed another string_view printf error caught by linux compiler

+++ Gems/ScriptCanvas/Code/Editor/View/Windows/MainWindow.cpp
@@ -1571,7 +1571,8 @@ namespace ScriptCanvasEditor
         
         if (outTabIndex == -1)
         {
-            return AZ::Failure(AZStd::string::format("Script Canvas Asset %.*s is not open in a tab", assetPath.size(), assetPath.data()));
+            return AZ::Failure(AZStd::string::format("Script Canvas Asset %.*s is not open in a tab"
+                , static_cast<int>(assetPath.size()), assetPath.data()));
         }
 
         SetActiveAsset(handle);
